fix assign task logic #426; create delete Template logic #423

+++ staff/miguel-arias/project/app/src/components/Template.jsx
@@ -1,17 +1,22 @@
 import { Button } from '../library'
 
+import { useContext } from '../hooks'
+
 import logic from '../logic'
 import helper from '../logic/helpers'
 
 function Template(props) {
     const template = props.template
+    const context = useContext()
 
     const handleDeleteClick = async () => {
-        try {
-            await logic.deleteTemplate(template.id)
-        } catch (error) {
-
-        }
+        if (confirm('Are you sure you want to delete this template? All tasks related to this will be deleted too'))
+            try {
+                await logic.deleteTemplate(template.id)
+                props.onDeleteSuccess()
+            } catch (error) {
+                context.handleError(error)
+            }
     }
 
     return <article>
@@ -23,7 +28,5 @@ function Template(props) {
         </aside>}
     </article>
 }
-
-/*  */
 
 export default Template
